requirements.txt
crewai
langchain
streamlit
openai
python-dotenv
langchain-litellm